WIP -Added PostRegister + File Clean up
Post Register is handles the creation of different post types, it allows
handles the generation of adding types to the sidebar in the admin
section.

Even the default page post is using it.

`
PostRegister::register([
	'name' => 'page',
        'urlBase' => '/',
]);

`

App Namespace changed. Theme facade added as a WIP to handle management of
assets and template checks.